Game

Description

Documentation

data GameState

Represents current game state

Constructors

GameState

Fields

_character :: Character

_monsterHP :: Int

character :: Lens' GameState Character

monsterHP :: Lens' GameState Int

valInRange :: Int -> Int -> IO Int

Used for calculating random damage or hp

initializeMonsterHP :: GameState -> Monster -> IO GameState

Takes current gamestate (monsterHP set to 0 or undefined), returns updated gamestate with calculated monsterHP

playerCommand :: GameState -> Monster -> IO GameState

Parsing and applying command ( ATTACK OR RUN )

attack :: GameState -> Monster -> IO GameState

'Runs' attack command

run :: GameState -> Monster -> IO GameState

'Runs' run command

encounter :: GameState -> Monster -> IO GameState

Used for setting combat

initializePlayer :: IO GameState

Initializes player (calls Character.playerCharacter) and returns updated Gamestate

gauntlet :: Int -> GameState -> IO (Int, GameState)

Game loop until player dies or runs. Keeps track of # of monsters killed

loop :: IO GameState

Runs game loop, prints record to file, sort records according to # of kills, returns final gamestate

formatRecord :: GameState -> Int -> String

Format printed to text file

parseRecords :: String -> Maybe (String, CharacterClass, Int)

parseRecord extracts name, class, and kills from the input string (to be used in a line from file)

writeRecord :: FilePath -> GameState -> Int -> IO ()

Writes formated record to file

readAndSortRecords :: FilePath -> IO ()

Read and sort records based on kills

formatSortedRecord :: (String, CharacterClass, Int) -> String

Convert parsed records (String, CharacterClass, Int) into GameState and apply formatRecord